refactor/3 ENUM 클래스 패키지명 수정 - enumeration -> model
