+++ src/main/java/kr/gravy/gravy/auth/jwt/JWTAuthenticationFilter.java
@@ -7,7 +7,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import kr.gravy.gravy.common.exception.GravyException;
 import kr.gravy.gravy.common.exception.Status;
-import kr.gravy.gravy.auth.enumeration.Grade;
+import kr.gravy.gravy.auth.model.Grade;
 import kr.gravy.gravy.auth.mapper.UserMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
