+++ src/main/java/kr/gravy/gravy/auth/service/AuthService.java
@@ -3,7 +3,7 @@ package kr.gravy.gravy.auth.service;
 import kr.gravy.gravy.auth.dto.ReIssueAccessTokenDto;
 import kr.gravy.gravy.auth.dto.UserLoginDto;
 import kr.gravy.gravy.auth.dto.UserSignUpDto;
-import kr.gravy.gravy.auth.enumeration.Grade;
+import kr.gravy.gravy.auth.model.Grade;
 import kr.gravy.gravy.auth.jwt.JWTUtil;
 import kr.gravy.gravy.auth.mapper.RefreshTokenMapper;
 import kr.gravy.gravy.auth.mapper.UserMapper;
@@ -12,7 +12,7 @@ import kr.gravy.gravy.auth.vo.UserVO;
 import kr.gravy.gravy.common.exception.GravyException;
 import kr.gravy.gravy.common.exception.Status;
 import kr.gravy.gravy.common.utils.GeneratorUtil;
-import kr.gravy.gravy.email.enumeration.EmailVerificationStatus;
+import kr.gravy.gravy.email.model.EmailVerificationStatus;
 import kr.gravy.gravy.email.mapper.EmailVerificationMapper;
 import kr.gravy.gravy.email.vo.RefreshTokenVO;
 import lombok.RequiredArgsConstructor;
