+++ src/main/java/kr/gravy/gravy/email/model/EmailVerificationStatus.java
@@ -1,4 +1,4 @@
-package kr.gravy.gravy.email.enumeration;
+package kr.gravy.gravy.email.model;
 
 public enum EmailVerificationStatus {
     SENT,
